Corrigir atualizacao das atividades
